requirements.txt
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
dnspython==2.7.0
dotenv==0.9.9
Flask==3.1.1
Flask-JWT-Extended==4.7.1
Flask-PyMongo==3.0.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
PyJWT==2.10.1
pymongo==4.13.0
python-dotenv==1.1.0
requests==2.32.3
RouterOS-api==0.21.0
urllib3==2.4.0
Werkzeug==3.1.3
